refactor: melhora os nomes
